requirements.txt
opencv-python-headless==4.7.0.72
Flask==2.0.3
gTTS==2.3.1
playsound==1.3.0
pillow==9.1.0
pytesseract==0.3.9
Jinja2==3.0.3
requests<3
requests-cache==0.9.8
Werkzeug==2.0.3
gunicorn==20.1.0
numpy<2
